Fix EBS volume cleanup partly out of my own sillyness and partly because of bugs

+++ cookbooks/ec-harness/recipes/cleanup.rb
@@ -20,30 +20,34 @@ if node['harness']['provider'] == 'ec2' && node['harness']['ec2']['backend_stora
         Chef::Log.info("Databag ebs_volumes_db not found, continuing")
       end
 
-      # destroy the volume (scary!)
-      fog_helper = FogHelper.new(region: node['harness']['ec2']['region'])
-      begin
-        fog_helper.get_aws.delete_volume(ebs_vol_id)
-      rescue Fog::Compute::AWS::Error => e
-        # typically due to volume still attached, machine shutting down
-        Chef::Log.info("AWS error, sleeping and then retrying: #{e}")
-        sleep 5
-        retry
-      rescue Fog::Compute::AWS::NotFound => e
-        # because the volume doesn't exist
-        Chef::Log.info("Volume not found, continuing")
+      if ebs_vol_id
+        fog_helper = FogHelper.new(region: node['harness']['ec2']['region'])
+        begin
+          fog_helper.get_aws.delete_volume(ebs_vol_id)
+        rescue Fog::Compute::AWS::Error => e
+          # typically due to volume still attached, machine shutting down
+          Chef::Log.info("AWS error, sleeping and then retrying: #{e}")
+          sleep 5
+          retry
+        rescue Fog::Compute::AWS::NotFound
+          # because the volume doesn't exist
+          Chef::Log.info("Volume not found, continuing")
+        end
       end
     end
   end
 
   ruby_block 'delete_ebs_volumes_databag_item' do
     block do
-      topology = TopoHelper.new(ec_config: node['harness']['vm_config'])
       begin
-        item = data_bag_item('ebs_volumes_db', topology.bootstrap_host_name)
-        item.destroy
-      rescue Net::HTTPServerException
-        Chef::Log.info("Databag ebs_volumes_db not found, continuing")
+        ebs_volumes_db = Chef::DataBag.new
+        ebs_volumes_db.name('ebs_volumes_db')
+        ebs_volumes_db.destroy
+        # can't destroy an item because https://github.com/opscode/chef-zero/issues/82
+        # item = data_bag_item('ebs_volumes_db', topology.bootstrap_host_name)
+        # item.destroy
+      rescue Net::HTTPServerException => e
+        Chef::Log.info("Databag ebs_volumes_db not found, continuing: #{e}")
       end
     end
   end
